+++ motionStudio/src/engines/canvas/index.ts
@@ -1,3 +1,3 @@
 export { useCanvasEngine } from './store';
 export type { ElementPatch } from './store';
-export type { CanvasElement, TextElement, ImageElement, VideoElement, AudioElement, BaseElement, AddTextInput } from './types';
+export type { CanvasElement, TextElement, ImageElement, VideoElement, AudioElement, ShaderElement, BaseElement, ShaderPreset, AddTextInput } from './types';